// web/src/lib/types.ts
export type DockerContainerStats = {
  id: string;
  name: string;
  cpu_percent: number;
  mem_bytes: number;
  mem_limit_bytes: number;
  net_rx_bytes: number;
  net_tx_bytes: number;
  blk_read_bytes: number;
  blk_write_bytes: number;
  pids: number;
};